+++ examples/resources/iosxe_bgp/resource.tf
@@ -2,6 +2,7 @@ resource "iosxe_bgp" "example" {
   asn                  = "65000"
   default_ipv4_unicast = false
   log_neighbor_changes = true
+  asnotation_dot       = true
   bgp_graceful_restart = true
   bgp_update_delay     = 200
 }