+++ internal/provider/data_source_iosxe_bgp_test.go
@@ -34,6 +34,7 @@ func TestAccDataSourceIosxeBGP(t *testing.T) {
 	var checks []resource.TestCheckFunc
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp.test", "default_ipv4_unicast", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp.test", "log_neighbor_changes", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp.test", "asnotation_dot", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp.test", "bgp_graceful_restart", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_bgp.test", "bgp_update_delay", "200"))
 	resource.Test(t, resource.TestCase{
@@ -73,6 +74,7 @@ func testAccDataSourceIosxeBGPConfig() string {
 	config += `	asn = "65000"` + "\n"
 	config += `	default_ipv4_unicast = false` + "\n"
 	config += `	log_neighbor_changes = true` + "\n"
+	config += `	asnotation_dot = true` + "\n"
 	config += `	bgp_graceful_restart = true` + "\n"
 	config += `	bgp_update_delay = 200` + "\n"
 	config += `	depends_on = [iosxe_yang.PreReq0, ]` + "\n"