+++ internal/provider/model_iosxe_bgp.go
@@ -45,6 +45,7 @@ type BGP struct {
 	LogNeighborChanges types.Bool   `tfsdk:"log_neighbor_changes"`
 	RouterIdLoopback   types.Int64  `tfsdk:"router_id_loopback"`
 	RouterIdIp         types.String `tfsdk:"router_id_ip"`
+	AsnotationDot      types.Bool   `tfsdk:"asnotation_dot"`
 	BgpGracefulRestart types.Bool   `tfsdk:"bgp_graceful_restart"`
 	BgpUpdateDelay     types.Int64  `tfsdk:"bgp_update_delay"`
 }
@@ -57,6 +58,7 @@ type BGPData struct {
 	LogNeighborChanges types.Bool   `tfsdk:"log_neighbor_changes"`
 	RouterIdLoopback   types.Int64  `tfsdk:"router_id_loopback"`
 	RouterIdIp         types.String `tfsdk:"router_id_ip"`
+	AsnotationDot      types.Bool   `tfsdk:"asnotation_dot"`
 	BgpGracefulRestart types.Bool   `tfsdk:"bgp_graceful_restart"`
 	BgpUpdateDelay     types.Int64  `tfsdk:"bgp_update_delay"`
 }
@@ -106,6 +108,13 @@ func (data BGP) toBodyXML(ctx context.Context, config BGP) string {
 	}
 	if !data.RouterIdIp.IsNull() && !data.RouterIdIp.IsUnknown() {
 		body = helpers.SetFromXPath(body, data.getXPath()+"/bgp/router-id/ip-id", data.RouterIdIp.ValueString())
+	}
+	if !data.AsnotationDot.IsNull() && !data.AsnotationDot.IsUnknown() {
+		if data.AsnotationDot.ValueBool() {
+			body = helpers.SetFromXPath(body, data.getXPath()+"/bgp/asnotation/dot", "")
+		} else {
+			body = helpers.RemoveFromXPath(body, data.getXPath()+"/bgp/asnotation/dot")
+		}
 	}
 	if !data.BgpGracefulRestart.IsNull() && !data.BgpGracefulRestart.IsUnknown() {
 		if data.BgpGracefulRestart.ValueBool() {
@@ -158,6 +167,15 @@ func (data *BGP) updateFromBodyXML(ctx context.Context, res xmldot.Result) {
 	} else {
 		data.RouterIdIp = types.StringNull()
 	}
+	if value := helpers.GetFromXPath(res, "data"+data.getXPath()+"/bgp/asnotation/dot"); !data.AsnotationDot.IsNull() {
+		if value.Exists() {
+			data.AsnotationDot = types.BoolValue(true)
+		} else {
+			data.AsnotationDot = types.BoolValue(false)
+		}
+	} else {
+		data.AsnotationDot = types.BoolNull()
+	}
 	if value := helpers.GetFromXPath(res, "data"+data.getXPath()+"/bgp/gr-options/graceful-restart"); !data.BgpGracefulRestart.IsNull() {
 		if value.Exists() {
 			data.BgpGracefulRestart = types.BoolValue(true)
@@ -195,6 +213,11 @@ func (data *BGP) fromBodyXML(ctx context.Context, res xmldot.Result) {
 	if value := helpers.GetFromXPath(res, "data"+data.getXPath()+"/bgp/router-id/ip-id"); value.Exists() {
 		data.RouterIdIp = types.StringValue(value.String())
 	}
+	if value := helpers.GetFromXPath(res, "data"+data.getXPath()+"/bgp/asnotation/dot"); value.Exists() {
+		data.AsnotationDot = types.BoolValue(true)
+	} else {
+		data.AsnotationDot = types.BoolValue(false)
+	}
 	if value := helpers.GetFromXPath(res, "data"+data.getXPath()+"/bgp/gr-options/graceful-restart"); value.Exists() {
 		data.BgpGracefulRestart = types.BoolValue(true)
 	} else {
@@ -226,6 +249,11 @@ func (data *BGPData) fromBodyXML(ctx context.Context, res xmldot.Result) {
 	if value := helpers.GetFromXPath(res, "data"+data.getXPath()+"/bgp/router-id/ip-id"); value.Exists() {
 		data.RouterIdIp = types.StringValue(value.String())
 	}
+	if value := helpers.GetFromXPath(res, "data"+data.getXPath()+"/bgp/asnotation/dot"); value.Exists() {
+		data.AsnotationDot = types.BoolValue(true)
+	} else {
+		data.AsnotationDot = types.BoolValue(false)
+	}
 	if value := helpers.GetFromXPath(res, "data"+data.getXPath()+"/bgp/gr-options/graceful-restart"); value.Exists() {
 		data.BgpGracefulRestart = types.BoolValue(true)
 	} else {
@@ -247,6 +275,9 @@ func (data *BGP) addDeletedItemsXML(ctx context.Context, state BGP, body string)
 	}
 	if !state.BgpGracefulRestart.IsNull() && data.BgpGracefulRestart.IsNull() {
 		b = helpers.RemoveFromXPath(b, state.getXPath()+"/bgp/gr-options/graceful-restart")
+	}
+	if !state.AsnotationDot.IsNull() && data.AsnotationDot.IsNull() {
+		b = helpers.RemoveFromXPath(b, state.getXPath()+"/bgp/asnotation/dot")
 	}
 	if !state.RouterIdIp.IsNull() && data.RouterIdIp.IsNull() {
 		b = helpers.RemoveFromXPath(b, state.getXPath()+"/bgp/router-id/ip-id")
@@ -276,6 +307,9 @@ func (data *BGP) addDeletePathsXML(ctx context.Context, body string) string {
 	}
 	if !data.BgpGracefulRestart.IsNull() {
 		b = helpers.RemoveFromXPath(b, data.getXPath()+"/bgp/gr-options/graceful-restart")
+	}
+	if !data.AsnotationDot.IsNull() {
+		b = helpers.RemoveFromXPath(b, data.getXPath()+"/bgp/asnotation/dot")
 	}
 	if !data.RouterIdIp.IsNull() {
 		b = helpers.RemoveFromXPath(b, data.getXPath()+"/bgp/router-id/ip-id")