+++ internal/provider/resource_iosxe_bgp.go
@@ -115,6 +115,10 @@ func (r *BGPResource) Schema(ctx context.Context, req resource.SchemaRequest, re
 					stringvalidator.RegexMatches(regexp.MustCompile(`(([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])(%[\p{N}\p{L}]+)?`), ""),
 				},
 			},
+			"asnotation_dot": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("AS dot notation").String,
+				Optional:            true,
+			},
 			"bgp_graceful_restart": schema.BoolAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("Graceful restart capability parameters").String,
 				Optional:            true,