+++ internal/provider/resource_iosxe_bgp_test.go
@@ -37,6 +37,7 @@ func TestAccIosxeBGP(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp.test", "asn", "65000"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp.test", "default_ipv4_unicast", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp.test", "log_neighbor_changes", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp.test", "asnotation_dot", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp.test", "bgp_graceful_restart", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp.test", "bgp_update_delay", "200"))
 	resource.Test(t, resource.TestCase{
@@ -111,6 +112,7 @@ func testAccIosxeBGPConfig_all() string {
 	config += `	asn = "65000"` + "\n"
 	config += `	default_ipv4_unicast = false` + "\n"
 	config += `	log_neighbor_changes = true` + "\n"
+	config += `	asnotation_dot = true` + "\n"
 	config += `	bgp_graceful_restart = true` + "\n"
 	config += `	bgp_update_delay = 200` + "\n"
 	config += `	depends_on = [iosxe_yang.PreReq0, ]` + "\n"